requirements.txt
streamlit>=1.38.0
pydantic>=2.9.0
python-dotenv>=1.0.1
langchain>=0.3.0
chromadb>=0.5.5
beautifulsoup4>=4.12.3
requests>=2.32.3
tqdm>=4.66.4
google-cloud-aiplatform>=1.74.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pydantic[email]
pypdf
urllib